requirements.txt
PyQt5==5.15.9
python-snap7==1.3
